restructured fedsim module
