+++ build_all.sh
@@ -1,6 +1,5 @@
 #!/bin/sh
 # black --line-length 79 .
 isort .
-python setup.py bdist_wheel
-python setup.py sdist
+python setup.py bdist_wheel sdist
