+++ docs/source/conf.py
@@ -17,6 +17,9 @@ import sys
 
 sys.path.insert(0, os.path.abspath("../.."))
 
+import ipdb
+ipdb.set_trace()
+
 # -- Project information -----------------------------------------------------
 
 master_doc = "index"
@@ -71,7 +74,7 @@ html_theme = "pydata_sphinx_theme"
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ["_static"]
+# html_static_path = ["_static"]
 html_use_smartypants = True
 html_last_updated_fmt = "%b %d, %Y"
