--- docs/source/fedsim.data_manager.rst
@@ -1,37 +1,0 @@
-fedsim.data\_manager package
-============================
-
-Submodules
-----------
-
-fedsim.data\_manager.basic\_data\_manager module
-------------------------------------------------
-
-.. automodule:: fedsim.data_manager.basic_data_manager
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-fedsim.data\_manager.data\_manager module
------------------------------------------
-
-.. automodule:: fedsim.data_manager.data_manager
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-fedsim.data\_manager.utils module
----------------------------------
-
-.. automodule:: fedsim.data_manager.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: fedsim.data_manager
-   :members:
-   :undoc-members:
-   :show-inheritance:
