+++ docs/source/fedsim.distributed.centralized.compression.rst
@@ -1,0 +1,10 @@
+fedsim.distributed.centralized.compression package
+==================================================
+
+Module contents
+---------------
+
+.. automodule:: fedsim.distributed.centralized.compression
+   :members:
+   :undoc-members:
+   :show-inheritance:
