+++ docs/source/fedsim.distributed.centralized.privacy.rst
@@ -1,0 +1,10 @@
+fedsim.distributed.centralized.privacy package
+==============================================
+
+Module contents
+---------------
+
+.. automodule:: fedsim.distributed.centralized.privacy
+   :members:
+   :undoc-members:
+   :show-inheritance:
