+++ docs/source/fedsim.distributed.centralized.rst
@@ -1,0 +1,31 @@
+fedsim.distributed.centralized package
+======================================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   fedsim.distributed.centralized.compression
+   fedsim.distributed.centralized.privacy
+   fedsim.distributed.centralized.training
+
+Submodules
+----------
+
+fedsim.distributed.centralized.centralized\_fl\_algorithm module
+----------------------------------------------------------------
+
+.. automodule:: fedsim.distributed.centralized.centralized_fl_algorithm
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: fedsim.distributed.centralized
+   :members:
+   :undoc-members:
+   :show-inheritance:
