+++ docs/source/fedsim.distributed.centralized.training.rst
@@ -1,0 +1,61 @@
+fedsim.distributed.centralized.training package
+===============================================
+
+Submodules
+----------
+
+fedsim.distributed.centralized.training.adabest module
+------------------------------------------------------
+
+.. automodule:: fedsim.distributed.centralized.training.adabest
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+fedsim.distributed.centralized.training.fedavg module
+-----------------------------------------------------
+
+.. automodule:: fedsim.distributed.centralized.training.fedavg
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+fedsim.distributed.centralized.training.fedavgm module
+------------------------------------------------------
+
+.. automodule:: fedsim.distributed.centralized.training.fedavgm
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+fedsim.distributed.centralized.training.feddyn module
+-----------------------------------------------------
+
+.. automodule:: fedsim.distributed.centralized.training.feddyn
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+fedsim.distributed.centralized.training.fednova module
+------------------------------------------------------
+
+.. automodule:: fedsim.distributed.centralized.training.fednova
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+fedsim.distributed.centralized.training.fedprox module
+------------------------------------------------------
+
+.. automodule:: fedsim.distributed.centralized.training.fedprox
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: fedsim.distributed.centralized.training
+   :members:
+   :undoc-members:
+   :show-inheritance:
