+++ docs/source/fedsim.distributed.data_management.rst
@@ -1,0 +1,37 @@
+fedsim.distributed.data\_management package
+===========================================
+
+Submodules
+----------
+
+fedsim.distributed.data\_management.basic\_data\_manager module
+---------------------------------------------------------------
+
+.. automodule:: fedsim.distributed.data_management.basic_data_manager
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+fedsim.distributed.data\_management.data\_manager module
+--------------------------------------------------------
+
+.. automodule:: fedsim.distributed.data_management.data_manager
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+fedsim.distributed.data\_management.utils module
+------------------------------------------------
+
+.. automodule:: fedsim.distributed.data_management.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: fedsim.distributed.data_management
+   :members:
+   :undoc-members:
+   :show-inheritance:
