+++ docs/source/fedsim.distributed.decentralized.rst
@@ -1,0 +1,10 @@
+fedsim.distributed.decentralized package
+========================================
+
+Module contents
+---------------
+
+.. automodule:: fedsim.distributed.decentralized
+   :members:
+   :undoc-members:
+   :show-inheritance:
