+++ docs/source/fedsim.distributed.rst
@@ -1,0 +1,20 @@
+fedsim.distributed package
+==========================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   fedsim.distributed.centralized
+   fedsim.distributed.data_management
+   fedsim.distributed.decentralized
+
+Module contents
+---------------
+
+.. automodule:: fedsim.distributed
+   :members:
+   :undoc-members:
+   :show-inheritance:
