--- docs/source/fedsim.fl.algorithms.rst
@@ -1,61 +1,0 @@
-fedsim.fl.algorithms package
-============================
-
-Submodules
-----------
-
-fedsim.fl.algorithms.adabest module
------------------------------------
-
-.. automodule:: fedsim.fl.algorithms.adabest
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-fedsim.fl.algorithms.fedavg module
-----------------------------------
-
-.. automodule:: fedsim.fl.algorithms.fedavg
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-fedsim.fl.algorithms.fedavgm module
------------------------------------
-
-.. automodule:: fedsim.fl.algorithms.fedavgm
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-fedsim.fl.algorithms.feddyn module
-----------------------------------
-
-.. automodule:: fedsim.fl.algorithms.feddyn
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-fedsim.fl.algorithms.fednova module
------------------------------------
-
-.. automodule:: fedsim.fl.algorithms.fednova
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-fedsim.fl.algorithms.fedprox module
------------------------------------
-
-.. automodule:: fedsim.fl.algorithms.fedprox
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: fedsim.fl.algorithms
-   :members:
-   :undoc-members:
-   :show-inheritance:
