--- docs/source/fedsim.fl.mtl_algorithms.rst
@@ -1,10 +1,0 @@
-fedsim.fl.mtl\_algorithms package
-=================================
-
-Module contents
----------------
-
-.. automodule:: fedsim.fl.mtl_algorithms
-   :members:
-   :undoc-members:
-   :show-inheritance:
