--- docs/source/fedsim.fl.rst
@@ -1,54 +1,0 @@
-fedsim.fl package
-=================
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   fedsim.fl.algorithms
-   fedsim.fl.mtl_algorithms
-
-Submodules
-----------
-
-fedsim.fl.aggregators module
-----------------------------
-
-.. automodule:: fedsim.fl.aggregators
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-fedsim.fl.evaluation module
----------------------------
-
-.. automodule:: fedsim.fl.evaluation
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-fedsim.fl.fl\_algorithm module
-------------------------------
-
-.. automodule:: fedsim.fl.fl_algorithm
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-fedsim.fl.utils module
-----------------------
-
-.. automodule:: fedsim.fl.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: fedsim.fl
-   :members:
-   :undoc-members:
-   :show-inheritance:
