+++ docs/source/fedsim.local.training.rst
@@ -1,0 +1,21 @@
+fedsim.local.training package
+=============================
+
+Submodules
+----------
+
+fedsim.local.training.step\_closures module
+-------------------------------------------
+
+.. automodule:: fedsim.local.training.step_closures
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: fedsim.local.training
+   :members:
+   :undoc-members:
+   :show-inheritance:
