+++ docs/source/fedsim.utils.rst
@@ -1,0 +1,45 @@
+fedsim.utils package
+====================
+
+Submodules
+----------
+
+fedsim.utils.aggregators module
+-------------------------------
+
+.. automodule:: fedsim.utils.aggregators
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+fedsim.utils.convert\_parameters module
+---------------------------------------
+
+.. automodule:: fedsim.utils.convert_parameters
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+fedsim.utils.metric\_scores module
+----------------------------------
+
+.. automodule:: fedsim.utils.metric_scores
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+fedsim.utils.utils module
+-------------------------
+
+.. automodule:: fedsim.utils.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: fedsim.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
