+++ docs/source/modules.rst
@@ -5,3 +5,4 @@ fedsim
    :maxdepth: 4
 
    fedsim
+   scripts
