+++ docs/source/scripts.rst
@@ -1,0 +1,29 @@
+scripts package
+===============
+
+Submodules
+----------
+
+scripts.fed\_learn module
+-------------------------
+
+.. automodule:: scripts.fed_learn
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+scripts.simulate module
+-----------------------
+
+.. automodule:: scripts.simulate
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: scripts
+   :members:
+   :undoc-members:
+   :show-inheritance:
