+++ fedsim/distributed/centralized/training/adabest.py
@@ -1,21 +1,25 @@
 r""" This file contains an implementation of the following paper:
-    Title: "Federated Learning Based on Dynamic Regularization"
-    Authors: Durmus Alp Emre Acar, Yue Zhao, Ramon Matas, Matthew Mattina,
-    ----- Paul Whatmough, Venkatesh Saligrama
-    Publication date: [28 Sept 2020 (modified: 25 Mar 2021)]
-    Link: https://openreview.net/forum?id=B7v4QMR6Z9w
+    Title: "Minimizing Client Drift in Federated Learning via Adaptive
+    ---- Bias Estimation"
+    Authors: Farshid Varno, Marzie Saghayi, Laya Rafiee, Sharut Gupta,
+        Stan Matwin, Mohammad Havaei
+    Publication date: [Submitted on 27 Apr 2022 (v1), last revised
+    ---- 23 May 2022 (this version, v2)]
+    Link: https://arxiv.org/abs/2204.13170
 """
 from functools import partial
 
 import torch
 from torch.nn.utils import parameters_to_vector
 
-from ..utils import default_closure
-from ..utils import vector_to_parameters_like
+from fedsim.local.training.step_closures import default_closure
+from fedsim.utils import SerialAggregator
+from fedsim.utils import vector_to_parameters_like
+
 from . import fedavg
 
 
-class FedDyn(fedavg.FedAvg):
+class AdaBest(fedavg.FedAvg):
     def __init__(
         self,
         data_manager,
@@ -38,12 +42,17 @@ class FedDyn(fedavg.FedAvg):
         device="cuda",
         log_freq=10,
         mu=0.02,
+        beta=0.98,
         *args,
         **kwargs,
     ):
         self.mu = mu
+        self.beta = beta
+        # this is to avoid violations like reading oracle info and
+        # number of clients in FedDyn and SCAFFOLD
+        self.general_agg = SerialAggregator()
 
-        super(FedDyn, self).__init__(
+        super(AdaBest, self).__init__(
             data_manager,
             metric_logger,
             num_clients,
@@ -67,12 +76,11 @@ class FedDyn(fedavg.FedAvg):
 
         cloud_params = self.read_server("cloud_params")
         self.write_server("avg_params", cloud_params.detach().clone())
-        self.write_server("h", torch.zeros_like(cloud_params))
+
         for client_id in range(num_clients):
             self.write_client(client_id, "h", torch.zeros_like(cloud_params))
-        # oracle read violation, num_clients read violation
-        average_sample = len(self.oracle_dataset["train"]) / self.num_clients
-        self.write_server("average_sample", average_sample)
+            self.write_client(client_id, "last_round", -1)
+        self.write_server("average_sample", 0)
 
     def send_to_server(
         self,
@@ -98,8 +106,7 @@ class FedDyn(fedavg.FedAvg):
         )
 
         def transform_grads_fn(model):
-            params = parameters_to_vector(model.parameters())
-            grad_additive = 0.5 * (params - params_init) - h
+            grad_additive = -h
             grad_additive_list = vector_to_parameters_like(
                 mu_adaptive * grad_additive, model.parameters()
             )
@@ -110,7 +117,7 @@ class FedDyn(fedavg.FedAvg):
         step_closure_ = partial(
             default_closure, transform_grads=transform_grads_fn
         )
-        opt_res = super(FedDyn, self).send_to_server(
+        opt_res = super(AdaBest, self).send_to_server(
             client_id,
             datasets,
             epochs,
@@ -130,24 +137,29 @@ class FedDyn(fedavg.FedAvg):
             params_init
             - parameters_to_vector(model.parameters()).detach().clone().data
         )
-        new_h = h + pseudo_grads
+        t = self.rounds
+        new_h = (
+            1 / (t - self.read_client(client_id, "last_round")) * h
+            + pseudo_grads
+        )
         self.write_client(client_id, "h", new_h)
+        self.write_client(client_id, "last_round", self.rounds)
         return opt_res
 
     def receive_from_client(self, client_id, client_msg, aggregation_results):
         weight = 1
         self.agg(client_id, client_msg, aggregation_results, weight=weight)
+        self.general_agg.add(
+            "avg_m", client_msg["num_samples"] / self.epochs, 1
+        )
+        self.write_server("average_sample", self.general_agg.get("avg_m"))
 
     def optimize(self, aggregator):
         if "local_params" in aggregator:
-            weight = aggregator.get_weight("local_params")
             param_avg = aggregator.pop("local_params")
             optimizer = self.read_server("optimizer")
             cloud_params = self.read_server("cloud_params")
-            pseudo_grads = cloud_params.data - param_avg
-            h = self.read_server("h")
-            # read total clients VIOLATION
-            h = h + weight / self.num_clients * pseudo_grads
+            h = self.beta * (self.read_server("avg_params") - param_avg)
             new_params = param_avg - h
             modified_pseudo_grads = cloud_params.data - new_params
             # update cloud params
@@ -155,9 +167,6 @@ class FedDyn(fedavg.FedAvg):
             cloud_params.grad = modified_pseudo_grads
             optimizer.step()
             self.write_server("avg_params", param_avg.detach().clone())
-            self.write_server("h", h.data)
-            # purge aggregated results
-            del param_avg
         return aggregator.pop_all()
 
     def deploy(self):
