+++ fedsim/local/training/__init__.py
@@ -1,0 +1,2 @@
+from ._inference import local_inference
+from ._training import local_train
