+++ fedsim/utils/__init__.py
@@ -1,0 +1,11 @@
+from ._convert_parameters import vector_to_parameters_like
+from ._dict_opts import add_dict_to_dict
+from ._dict_opts import add_in_dict
+from ._dict_opts import append_dict_to_dict
+from ._dict_opts import apply_on_dict
+from ._dict_opts import reduce_dict
+from ._importing import get_from_module
+from ._importing import search_in_submodules
+from ._metric_scores import get_metric_scores
+from ._seed import set_seed
+from .aggregators import SerialAggregator
