+++ fedsim/utils/_dict_opts.py
@@ -1,46 +1,3 @@
-import importlib
-import inspect
-import random
-
-import numpy as np
-import torch
-
-
-def get_from_module(module_name, submodule_name, class_name):
-    module = importlib.import_module(module_name)
-    if submodule_name in module.__all__:
-        sub_module = importlib.import_module(
-            "{}.{}".format(module_name, submodule_name)
-        )
-        return getattr(sub_module, class_name)
-    raise NotImplementedError
-
-
-def search_in_submodules(module_name, object_name):
-    module = importlib.import_module(module_name)
-    for submodule_name in module.__all__:
-        sub_module = importlib.import_module(
-            "{}.{}".format(module_name, submodule_name)
-        )
-        for name, obj in inspect.getmembers(
-            sub_module,
-            lambda x: inspect.isclass(x) or inspect.isfunction(x),
-        ):
-            if name == object_name:
-                return obj
-    return None
-
-
-def set_seed(seed, use_cuda) -> None:
-    np.random.seed(seed)
-    random.seed(seed)
-    torch.manual_seed(seed)
-
-    if use_cuda:
-        torch.backends.cudnn.deterministic = True
-        torch.cuda.manual_seed_all(seed)
-
-
 def append_dict_to_dict(new_dict, currecnt_dict=None):
     ans = verify_dict(currecnt_dict)
     for key, item in new_dict.items():
