+++ tests/test_fedsim.py
@@ -1,7 +1,7 @@
 from torch.utils.tensorboard import SummaryWriter
 
-from fedsim.data_manager.basic_data_manager import BasicDataManager
-from fedsim.fl.algorithms.fedavg import FedAvg
+from fedsim.distributed.centralized.training.fedavg import FedAvg
+from fedsim.distributed.data_management.basic_data_manager import BasicDataManager
 from fedsim.models.mcmahan_nets import cnn_cifar100
